Cursor Patch: Dedicated file & streamline edits
Cursor Patch: Move to dedicated file, clean up its code, and use same unpatch approach as the other patch methods
